firebase compatibility.
