+++ app/src/main/java/in/ac/nitc/eyyauto/Constants.java
@@ -1,7 +1,7 @@
 package in.ac.nitc.eyyauto;
 
 public final class Constants {
-    public static final String ROOT_PATH = "EyyAuto/";
+    public static final String ROOT_PATH = "/";
     public static final String USER_INFO_ROOT_PATH = ROOT_PATH + "Users/";
     public static final String DRIVER_INFO_ROOT_PATH = ROOT_PATH + "Drivers/";
     public static final String REQUEST_ROOT_PATH = ROOT_PATH + "Requests/";
